+++ src/GraphNode/GraphEvent.php
@@ -30,7 +30,7 @@ class GraphEvent extends GraphNode
     /**
      * @var array maps object key names to GraphNode types
      */
-    protected static $graphObjectMap = [
+    protected static $graphNodeMap = [
         'cover' => GraphCoverPhoto::class,
         'place' => GraphPage::class,
         'picture' => GraphPicture::class,