+++ src/GraphNode/GraphPage.php
@@ -30,7 +30,7 @@ class GraphPage extends GraphNode
     /**
      * @var array maps object key names to Graph object types
      */
-    protected static $graphObjectMap = [
+    protected static $graphNodeMap = [
         'best_page' => GraphPage::class,
         'global_brand_parent_page' => GraphPage::class,
         'location' => GraphLocation::class,